fix(maven) to build and test phase 6

+++ server/src/main/java/service/LogoutService.java
@@ -8,6 +8,9 @@ public class LogoutService {
     public LogoutService(DataAccess dataAccess) {this.dataAccess = dataAccess;}
 
     public void logoutUser(String authToken) throws DataAccessException {
+        if(!dataAccess.isValidAuth(authToken)){
+            throw new DataAccessException("Error: unauthorized - invalid or expired authToken");
+        }
         dataAccess.deleteAuthToken(authToken);
     }
 }
